requirements.txt
telethon
aiohttp
PyGithub
python-decouple